move visiting widgets to widgets folder
