+++ lib/ui/screen/sight_search_screen/sight_search_screen_widget.dart
@@ -4,7 +4,7 @@ import 'package:places/constants/app_strings.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/screen/sight_search_screen/widgets/search_hystory_list_view.dart';
 import 'package:places/ui/screen/sight_search_screen/widgets/searched_sights_list_view.dart';
-import 'package:places/ui/screen/visiting_screen/empty_list_page.dart';
+import 'package:places/ui/screen/visiting_screen/widgets/empty_list_page.dart';
 import 'package:places/ui/widgets/buttons/app_bar_back_button.dart';
 import 'package:places/ui/widgets/error_pages/network_error_page.dart';
 import 'package:places/ui/widgets/search_bar.dart';
