+++ lib/ui/screen/visiting_screen/widgets/visited_page.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:places/constants/app_strings.dart';
 import 'package:places/data/model/sights/sight_want_to_visit.dart';
 import 'package:places/styles/custom_icons.dart';
-import 'package:places/ui/screen/visiting_screen/draggable_sight_cards_list_view.dart';
-import 'package:places/ui/screen/visiting_screen/empty_list_page.dart';
+import 'package:places/ui/screen/visiting_screen/widgets/draggable_sight_cards_list_view.dart';
+import 'package:places/ui/screen/visiting_screen/widgets/empty_list_page.dart';
 import 'package:places/ui/widgets/sight_cards/visited_sight_card.dart';
 import 'package:provider/provider.dart';
 
-import '../../../blocs/visiting_bloc/visiting_bloc.dart';
+import '../../../../blocs/visiting_bloc/visiting_bloc.dart';
 
 class VisitedPage extends StatelessWidget {
   const VisitedPage(this.sights);
@@ -25,8 +25,7 @@ class VisitedPage extends StatelessWidget {
             sights,
             onReplaceCard: (fromIndex, toIndex) {
               context.read<VisitingBloc>().add(
-                    VisitingEvent.changeVisitedCardsSequences(
-                        fromIndex: fromIndex, toIndex: toIndex),
+                    VisitingEvent.changeVisitedCardsSequences(fromIndex: fromIndex, toIndex: toIndex),
                   );
             },
           );
