+++ lib/ui/widgets/error_pages/network_error_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:places/constants/app_strings.dart';
 import 'package:places/styles/custom_icons.dart';
-import 'package:places/ui/screen/visiting_screen/empty_list_page.dart';
+import 'package:places/ui/screen/visiting_screen/widgets/empty_list_page.dart';
 
 class NetworkErrorPage extends EmptyListPage {
   NetworkErrorPage({
